+++ src/PaymentTools/Model/ICommitmentItem.cs
@@ -1,0 +1,9 @@
+﻿namespace PaymentTools.Model
+{
+    public interface ICommitmentItem
+    {
+        public string Type { get; }
+
+        public decimal Amount { get; }
+    }
+}
