+++ src/PaymentTools/Model/TransactionType.cs
@@ -1,0 +1,7 @@
+﻿namespace PaymentTools.Model
+{
+    public class TransactionType
+    {
+        public int TypeId { get; set; }
+    }
+}
